Orden de archivos
Se crearon nuevas clases y se creo la carpeta resources para almacenar las imagenes a utilizar
